#33 - separated file operations and added output flag
