+++ outputs/json.go
@@ -11,7 +11,7 @@ func NewJsonOutput() *JsonOutput {
 	return &JsonOutput{}
 }
 
-func (t *JsonOutput) DisplayCommits(v interface{}) error {
+func (t *JsonOutput) DisplayCommits(fg *FileGenerator, v interface{}) error {
 	data, err := json.Marshal(v)
 	if err != nil {
 		return err
