+++ outputs/types.go
@@ -1,5 +1,5 @@
 package outputs
 
 type Output interface {
-	DisplayCommits(interface{}) error
+	DisplayCommits(*FileGenerator, interface{}) error
 }
